feat: refactor to interface > implementation
