+++ src/main/java/com/estebandev/minicloud/service/FileManagerService.java
@@ -3,310 +3,49 @@ package com.estebandev.minicloud.service;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.FileAlreadyExistsException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.attribute.FileTime;
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.UUID;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.FileSystemResource;
 import org.springframework.core.io.Resource;
-import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.estebandev.minicloud.entity.FileMetadata;
-import com.estebandev.minicloud.entity.User;
-import com.estebandev.minicloud.entity.UserMetadata;
 import com.estebandev.minicloud.service.exception.FileIsNotDirectoryException;
 import com.estebandev.minicloud.service.utils.FileData;
-import com.estebandev.minicloud.service.utils.FileManagerUtils;
 
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
+public interface FileManagerService {
+    List<FileData> listFiles(String path)
+            throws FileNotFoundException, FileIsNotDirectoryException, IOException;
 
-@Service
-@RequiredArgsConstructor
-@Setter
-public class FileManagerService {
-    Logger logger = LoggerFactory.getLogger(this.getClass());
-    private final UserService userService;
-    private final ReentrantReadWriteLock fileLock = new ReentrantReadWriteLock();
-    private final FileMetadataService fileMetadataService;
+    void makeDirectory(String dirPathString)
+            throws FileAlreadyExistsException, IOException;
 
-    @Value("${var.filepath}")
-    private String pathString;
+    void makeDirectory(String dirPathString, String dirName)
+            throws FileAlreadyExistsException, IOException;
 
-    public void verifyRootDirectory() {
-        Path root = getRoot();
-        if (Files.exists(root) && Files.isDirectory(root) && Files.isWritable(root)) {
-            return;
-        }
+    void uploadFile(String dirPathString, MultipartFile multipartFile)
+            throws IOException, FileIsNotDirectoryException, FileNotFoundException;
 
-        try {
-            if (Files.exists(root) && !Files.isDirectory(root)) {
-                Files.delete(root);
-            }
+    Path rename(String pathString, String newName) throws IOException;
 
-            Files.createDirectory(root);
-        } catch (IOException e) {
-            logger.error("FATAL ERROR {}", e.getMessage());
-            System.exit(1);
-        }
-    }
+    Resource findFile(String pathString) throws FileNotFoundException, IOException;
 
-    // @PreAuthorize("@customFilesAuthenticationManager.check(#root)")
-    public List<FileData> listFiles(String path)
-            throws FileNotFoundException, FileIsNotDirectoryException, IOException {
-        fileLock.readLock().lock();
-        try {
+    FileData findFileData(String pathString) throws FileNotFoundException, IOException;
 
-            verifyRootDirectory();
+    public void delete(String pathString) throws FileNotFoundException, IOException;
 
-            Path root = getRoot();
-            Path dir = root.resolve(path);
+    Path getRoot();
 
-            if (!Files.exists(dir))
-                throw new FileNotFoundException();
+    long getLastModifiedDateInMinutes(String pathString) throws IOException;
 
-            if (!Files.isDirectory(dir))
-                throw new FileIsNotDirectoryException("The file is not directory");
+    boolean isExist(String path);
 
-            return Files.list(dir)
-                    .filter(filePath -> !filePath.getFileName().toString().startsWith(".dir"))
-                    .map(filePath -> {
-                        return FileData.builder()
-                                .fileName(filePath.getFileName().toString())
-                                .path(getRoot().relativize(filePath))
-                                .directory(Files.isDirectory(filePath))
-                                .build();
-                    })
-                    .toList();
-        } finally {
-            fileLock.readLock().unlock();
-        }
-    }
+    boolean isValidDirectory(String path);
 
-    @Transactional
-    private void makeDirectory(Path dirPath)
-            throws FileAlreadyExistsException, IOException {
-        fileLock.writeLock().lock();
-        try {
-            verifyRootDirectory();
+    String getMimeType(String filePathString) throws IOException;
 
-            if (Files.exists(dirPath))
-                throw new FileAlreadyExistsException("The already exist");
+    void savePathMetadata(Path dirPath) throws IOException;
 
-            Files.createDirectory(dirPath);
+    void uploadPathChild(Path parentPath) throws IOException;
 
-            fileMetadataService.make(dirPath, userService.getUserFromAuth());
-            savePathMetadata(dirPath);
-        } finally {
-            fileLock.writeLock().unlock();
-        }
-    }
-
-    public void makeDirectory(String dirPathString)
-            throws FileAlreadyExistsException, IOException {
-        Path dirPath = getRoot().resolve(dirPathString).normalize();
-
-        makeDirectory(dirPath);
-    }
-
-    public void makeDirectory(String dirPathString, String dirName)
-            throws FileAlreadyExistsException, IOException {
-        makeDirectory(getRoot().resolve(dirPathString).normalize().resolve(FileManagerUtils.formatName(dirName)));
-    }
-
-    public void uploadFile(String dirPathString, MultipartFile multipartFile)
-            throws IOException, FileIsNotDirectoryException, FileNotFoundException {
-        try {
-            if (!fileLock.writeLock().tryLock(5, TimeUnit.SECONDS)) {
-                throw new IOException("Upload is unable now");
-            }
-
-            try {
-                verifyRootDirectory();
-
-                Path dirPath = getRoot().resolve(dirPathString);
-
-                if (!Files.exists(dirPath))
-                    throw new FileNotFoundException("File does not exist");
-
-                if (!Files.isDirectory(dirPath))
-                    throw new FileIsNotDirectoryException(dirPath + " is not directory");
-
-                String fileName = FileManagerUtils.formatName(multipartFile.getOriginalFilename());
-                Path filePath = dirPath.resolve(fileName);
-                if (Files.exists(filePath)) {
-                    filePath = dirPath.resolve(FileManagerUtils.uniqueName(fileName));
-                }
-
-                multipartFile.transferTo(filePath);
-            } finally {
-                fileLock.writeLock().unlock();
-            }
-        } catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-            throw new IOException("Interrupt to try get lock", e);
-        }
-    }
-
-    @Transactional
-    public Path rename(String pathString, String newName) throws IOException {
-        try {
-            if (!fileLock.writeLock().tryLock(5, TimeUnit.SECONDS)) {
-                throw new IOException("Rename is unable now");
-            }
-
-            try {
-                Path filePath = getRoot().resolve(pathString);
-                Path newFilePath = filePath.getParent().resolve(FileManagerUtils.formatName(newName));
-
-                Files.move(filePath, newFilePath);
-
-                savePathMetadata(newFilePath);
-                return getRoot().relativize(newFilePath);
-            } finally {
-                fileLock.writeLock().unlock();
-            }
-        } catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-            throw new IOException("Interrupt to try get lock", e);
-        }
-    }
-
-    public Resource findFile(String pathString) throws FileNotFoundException, IOException {
-        fileLock.readLock().lock();
-        try {
-            verifyRootDirectory();
-
-            Path filePath = getRoot().resolve(pathString);
-
-            if (!Files.exists(filePath))
-                throw new FileNotFoundException("File not exists");
-
-            if (Files.isDirectory(filePath))
-                throw new IOException("The file is a directory");
-            if (filePath.getFileName().toString().contains(".dirdata"))
-                throw new IOException("The file isn't accesible");
-
-            FileManagerUtils.validateFile(filePath);
-
-            return new FileSystemResource(filePath);
-
-        } finally {
-            fileLock.readLock().unlock();
-        }
-    }
-
-    public FileData findFileData(String pathString) throws FileNotFoundException, IOException {
-        Path path = getRoot().resolve(pathString).normalize();
-        String fileName = path.getFileName().toString();
-        boolean editable = true;
-
-        if (userService.getUserFromAuth().getEmail().equalsIgnoreCase(fileName)
-                || path.equals(getRoot().normalize()))
-            editable = false;
-
-        return FileData.builder()
-                .fileName(fileName)
-                .path(path)
-                .mediaType(FileManagerUtils.getMimeType(path))
-                .size(FileManagerUtils.convertBytesToMegabytes(Files.size(path)))
-                .directory(Files.isDirectory(path))
-                .editable(editable)
-                .build();
-    }
-
-    public void delete(String pathString) throws FileNotFoundException, IOException {
-        try {
-            if (!fileLock.writeLock().tryLock(5, TimeUnit.SECONDS)) {
-                throw new IOException("Renable is unable now");
-            }
-            try {
-                Path filePath = getRoot().resolve(pathString);
-
-                if (!Files.exists(filePath))
-                    throw new FileNotFoundException();
-                if (!Files.isWritable(filePath))
-                    throw new IOException("You do not have perms");
-                if (Files.isDirectory(filePath) && !listFiles(pathString).isEmpty())
-                    throw new IOException("The directory is not empty");
-                if (Files.isDirectory(filePath))
-                    fileMetadataService.deleteAll(filePath);
-
-                Files.delete(filePath);
-            } finally {
-                fileLock.writeLock().unlock();
-            }
-        } catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-            throw new IOException("Interrupt to try get lock", e);
-        }
-    }
-
-    public Path getRoot() {
-        return Path.of(pathString).normalize();
-    }
-
-    public long getLastModifiedDateInMinutes(String pathString) throws IOException {
-        Path filePath = getRoot().resolve(pathString);
-        FileTime fTime = Files.getLastModifiedTime(filePath);
-        return fTime.to(TimeUnit.MINUTES);
-    }
-
-    public boolean isExist(String path) {
-        return Files.exists(getRoot().resolve(path));
-    }
-
-    public boolean isValidDirectory(String path) {
-        return isExist(path) && Files.isDirectory(getRoot().resolve(path));
-    }
-
-    public String getMimeType(String filePathString) throws IOException {
-        return FileManagerUtils.getMimeType(getRoot().resolve(filePathString));
-    }
-
-    public void savePathMetadata(Path dirPath) throws IOException {
-        if (!Files.isDirectory(dirPath))
-            return;
-
-        FileMetadata metadata;
-        try {
-            metadata = fileMetadataService.findMetadataFromKey(dirPath, "path");
-            metadata.setValue(getRoot().relativize(dirPath).toString());
-            uploadPathChild(dirPath);
-        } catch (NoSuchElementException e) {
-            metadata = FileMetadata.builder()
-                    .key("path")
-                    .value(getRoot().relativize(dirPath).toString())
-                    .build();
-        }
-
-        fileMetadataService.save(dirPath, metadata);
-    }
-
-    public void uploadPathChild(Path parentPath) throws IOException {
-        if (!Files.isDirectory(parentPath))
-            return;
-
-        Files.list(parentPath)
-                .filter(p -> Files.isDirectory(p))
-                .forEach(t -> {
-                    try {
-                        savePathMetadata(t);
-                    } catch (IOException e) {
-                        // Do nothing
-                    }
-                });
-    }
+    void setPathString(String path);
 }
